+++ src/lib/message.test.ts
@@ -21,7 +21,7 @@ describe('parseMessage', () => {
     expect(message.type).toBe('message')
     expect(message.timeStamp).toBe('1570172544.014600')
     expect(message.text).toBe(
-      '<@USERID> <#CHANNELID> Sample message\nSample\n\nSample'
+      '<@USERID> <#CHANNELID|general> Sample message\nSample\n\nSample'
     )
     expect(message.clientMessageId).toBe('363953ac-f635-4cfa-99f3-70f87b04a9cc')
     expect(message.team).toBe('TEAMID')